Removed session file extension 🥷🏻
