+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 1.9.4
 
+- Moved all methods in `SessionFileExtension` and `SessionsSaver` to `Session` and `SessionsManager` classes respectively.
+- This is to simplify the API and to make it easier to use.
+
+## 1.9.4
+
 - Added `SessionFileExtension` class to represent a file extension.
 - Added `SessionsSaver` extension on `SessionsManager` class to save all sessions to a directory.
 - Updated [/example/session_bot.dart](./example/session_bot.dart) file to reflect the changes.
