+++ lib/televerse.dart
@@ -44,7 +44,6 @@ part 'src/televerse/markups/keyboard.dart';
 part 'src/televerse/markups/inline_keyboard.dart';
 part 'src/televerse/raw_api.dart';
 part 'src/televerse/sessions/sessions.dart';
-part 'src/televerse/sessions/session_file_extension.dart';
 
 /// Type aliases for the library.
 part 'src/types/aliases.dart';
